+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Space` keyboard shortcut
 - `ArrowLeft` keyboard shortcut
 - `ArrowRight` keyboard shortcut
+- `ArrowUp` keyboard shortcut 
+- `ArrowDown` keyboard shortcut
 - `revokeCurrentMedia` function
 ### Changed
 - `README.md` updated